Replace base.py with __init__.py, expressions/base.py with expression/expressions.py
