+++ python/tests/conftest.py
@@ -32,7 +32,7 @@ from urllib.parse import urlparse
 import pytest
 
 from pyiceberg import schema
-from pyiceberg.io.base import (
+from pyiceberg.io import (
     FileIO,
     InputFile,
     OutputFile,
@@ -52,8 +52,8 @@ from pyiceberg.types import (
     StringType,
     StructType,
 )
-from tests.catalog.test_base import InMemoryCatalog
-from tests.io.test_io_base import LocalInputFile
+from tests.catalog.test_catalog import InMemoryCatalog
+from tests.io.test_io import LocalInputFile
 
 
 class FooStruct:
